Add pipeline components for running hhsuite tools
These components wrap the `phyre_engine.tools.hhsuite.*` tools for
running hhsuite tools. Only hhsearch and hhblits are supported so far.

These components do not parse output files, they simply add the
filenames to the data blob. It is the responsibility of the next
component in the pipeline to do the parsing.