Add some data files for testing purposes.
The "mmcif" data directory contains MMCIF files as downloaded from the
RCSB. The mmcif_label_only directory is for MMCIF files that have had
their "auth_..." fields removed, leaving only the "label_..." fields.
It is hoped that this will be useful in the future if the biopython
library catches up and is actually able to parse these files.