Sanitise template PDB files manually.
Previously, we used the Bio.PDB.PDBIO.Select class to control which
residues we output to our template files. This commit removes that from
phyre_engine.component.db.ChainPDBBuilder, so that we now sanitise our
PDBs manually.

This means that we can control the numbering of residues more easily,
and we can explicitly wipe insertion codes out. This also allows us to
build the mapping of residue index to author-assigned ID more easily.