Bugfix: Egregious error with disordered atoms
We needed to do a bit more work to handle disordered atoms. Rather than
checking if the conformation of each disordered atom is "A" and only
keeping those atoms, we explicitly choose conformation "A" for
disordered atoms, remembering to reset the "disordered_flag" so that
PDBIO does not try and write the atom as a disordered atom.