Treat PDBConstructionExceptions as warnings
At least for the moment, it shouldn't be considered a catastrophic
failure if an MMCIF file cannot be parsed (an altogether too common
occurrence), and so we catch and log PDBConstructionExceptions when they
occur.

All other exceptions are chained with a ChainExtractionError class that
provides useful info about the structure ID and chain. These remain
fatal.