Move AngleRange class to avoid circular dependency
My un-pythonic inclusion of class definitions in __init__.py is coming
back to bite me. In this case, there was a circular dependency when I
wanted to use "AngleRange" in the "data" module, and the "data" module
in __init__.py.

For the moment, I am just going to move AngleRange to its own module to
avoid this. In the future, after merging all development branches to the
master branch, I will refactor everything to bring class definitions out
of __init__.py files.