Update MolProbity rotamer definitions
The older definitions disagreed slightly with the output of the
"Rotalyze" class in chirporaxis.jar. The new version still disagrees
slightly, because MolProbity calculates the final chi angle of some PHE
and TYR residues as being greater than 180 degrees. I'm not sure how
this happens, or what the reasoning is, but it seems sufficiently rare
not to worry about it for the moment.

My suspicion is that this is due to a bug in MolProbity, but I can't be
sure because it's too much of a chore to learn how to operate Yet
Another Build System, compile and debug their code.