Remove code from __init__.py
It's not very pythonic to include lots of code in __init__.py files.
Lots of these changes are just replacing single-file packages with
modules.

Possibly in the future I will add import statements to __init__.py files
where only certain parts of the package ought to be in the public API.