Remove "stash" option to git
The "stash" option tried to do too much magic. Instead of stashing local
changes to a git repository, merging the master and unstashing changes,
we now just fetch the master and check for a diff. It will be the
responsibility of the build component to actually merge the remote
changes.