Refactor PBS classes
This commit is a big refactor of all the PBS code. It removes the idea
of a "resume" script, and instead adds job IDs to the pipeline. These
job IDs can then be processed in whichever way is most useful: we can
wait on them by polling qstat (using the Wait class), or detach (using
the Detatch) with a dependency on previous jobs.