Remove "AllMutationSelector" class
The "AllMutationSelector" class was intended to take every combination
of point mutations in a structure and generate a chain for each. This
doesn't really fit with our current data model, which expects a
one-to-one mapping of input chains to output chains.

It also isn't clear if the AllMutationSelector made sense: if residue 1
has mutations A and B, and residue 3 has A, B and C, how do we determine
what combinations to generate for conformation C?