Allow template grouping by sequence
This adds the phyre_engine.component.db.db.{Reduce,Expand} components.
These can be used to reduce the number of templates by representing all
templates with identical sequences with with a single template. The
Expand component can be used to convert this back to a flat list
containing all templates.