Completely refactor hhsuite report parser
This commit separates the ReportParser and TabularParser components, so
that the ReportParser can be used when hh{blits,search} was run without
the "-atab" switch.

The tools used for parsing the reports now return namedtuples rather
than dicts. This was felt to be a good idea because it more rigidly
structures the data extracted from each report. Namedtuples can be
trivially converted to dicts for use in components.