Use a bare string for the "sequence" field.
Having the "sequence" field point to a Bio.SeqRecord.SeqRecord object
was becoming cumbersome. It was difficult to serialise the pipeline, and
it made simple comparisons between sequences more difficult.

This commit replaces the "sequence" field with a plain string of
single-letter amino acid codes. The "phyre_engine.component.input"
components have been modified to return a "seq_record" key instead,
which contains a SeqRecord object.