Store chain metadata in template files
This commit is in preparation for the later inclusion of ligands in
models.

Previously, template chains were relatively fragile. The sequence of the
PDB was extracted directly from the ATOM records, meaning that template
files could only contain those residues that ought to appear in the
template HMM. This commit stores three pieces of metadata in the REMARK
fields of each template file:

1. The mapping between renumbered residues and the original residues IDs
   in the template PDB. This preserves arcane residue mappings,
   including insertion codes and HETATM flags.

2. The canonical sequence of the template file, extracted from all amino
   acids with ATOM records. Only the standard 20 amino acids are
   eligible for inclusion in the canonical sequence, and each included
   residue must contain a CA atom.

3. The list of renumbered residue IDs that are included in the canonical
   sequence. This will allow a homology modelling component to directly
   look up amino acids in an alignment without having to do a sequence
   alignment, while still retaining ligands or non-standard amino acids.