The "pbs" module now serialises pipeliens as JSON
Previously, we lumped the pipeline and the pipeline state together as a
single pickle file and loaded that in the remote job. This caused
problems: the pipeline state is supposed to be relatively simple, and so
could be serialised easily, but the pipeline itself could often not be
pickled.

This commit uses the Pipeline.load method to load a JSON representation
of the pipeline, so hopefully we can now avoid pickling complex data
types.

While we're at it, this commit normalises the parameter names for
components in the pbs package to be more in line with the SubPipeline
components. Rather than a "worker_pipeline" and "worker_config"
parameter, each component now accepts a "pipeline" parameter.