Remove CompletedState class.
Previously, the pipeline state was wrapped in
phyre_engine.component.pbs.qsub.CompletedState namedtuple. To verify
that the remote jobs were actually run, we would check whether the state
is a CompletedState.

This commit changes that to write the state and only the state to a
pickle. Our remote job runner adds a "qsub_completed" variable to the
state, which is checked and removed when it is loaded.