Update the .test.require* decorators
Rather than assign to the global phyre_engine.test.config variable, use
an "override_config" option to the requireField and requireConfig
decorators. This makes testing the decorators themselves significantly
easier, and prevents problems with subsequent tests relying on a broken
config variable.

I knew that overwriting a global variable in a test was (extremely)
sloppy, but I thought that the setUpModule and tearDownModule methods
would handle it. Obviously not.