Allow duplicate flags for external tools
If a list is passed as a value to the options parameter of an
ExternalTool, that option will now be duplicated. That is, passing
'options={"d": [1,2,3]}}' will now give the command line "-d 1 -d 2 -d
3".

This is in preparation for work with BackPhyre, when it will make life
much easier to pass multiple databases to hhblits.