Change output format for secstruc and disorder
The previous output format was rather opaque: just a tuple containing
the assigned state in the first field and an arbitrary bunch of
tool-specific keys. This was no good for showing on the front end.

The output format for these tools now includes the assigned state as a
single element and a separate element containing the confidence of each state.