If all our jobs disappear, assume they're done
We can't rely on the "C" state for qsub jobs to be assigned properly. If
the scheduler crashes (distressingly common) then the child jobs will
complete once the scheduler has restarted but the array job will not be
updated. It will eventually be purged, so we will assume that if we
don't see any of our target job IDs, all of them have finished.