Rename "TMAlignment" class to "TMAlign.Result"
I plan on including more classes in the phyre_engine.tools.strucaln
module, so to prepare I have made the "Result" class a nested class
underneath the "TMAlign" class. This opens the door to allowing, for
example, TMScore and TMScore.Result.

This commit also fixes some problems with the test suite, and moves over
to the @phyre_engine.test.requireFields method of skipping tests for
which the binaries cannot be found.