Fix up the LoopModeller component
Allow for alternate executables to be used (sbg-0.6.6 calls the loop
assembler "nova", while older versions called it "loop.assembler). This
commit also removes a nasty workaround that was required when running an
old version of the assembler on large file systems. Finally, it actually
updates the pipeline state with the new model.