Fix bugs in SoedingSelect: no longer greedy
This commit fixes some problems with the implementation of the Soeding
algorithm. Previously, the algorithm would accept the *first* template
with a global score greater than zero and then start again from the top
of the list. This is incorrect: templates should only be accepted when
all have been examined and the top-scoring template determined.