LoopModel component operates on a single model
This is part of a movement away from monolithic components that operate
on the entire pipeline state. Instead of operating on the "templates"
list, the LoopModel component will now operate on a single model at a
time, allowing the usual mechanisms (Map, TryCatch) to be used for
iterating over each model.

This commit also adds a test case for the loop modeller, because I was
apparaently in a big hurry when I wrote it.