Extremely breaking: Disallow positional args
This commit disallows positional arguments from pipeline specification
files. It is no longer possible to specify a component like so:

    - phyre_engine.component.foo.Bar: [a, b, c]

This syntax was functional, and it works well for trivial components,
but for complex pipelines or components it can quickly become
unmanageable. Does "Copy: [a, b]" mean to copy "a" into "b" or
vice-versa? Rather than try and evolve a set of conventions that I (and
any other eventual collaborators) will forget about, we will force the
pipeline specifications to be more explicit:

    - phyre_engine.component.foo.Bar:
        source: a
        destination: b
        threshold: c

This will increase the amount of boilerplate required in pipeline specs,
but hopefully increase the obviousness.