Improve error handling for JMESPath components
This adds checks (and tests for those checks) for the types of the
JMESPath results. We also add a check for an edge case in which the
result of select_expr is the same as the result of value_expr, which was
causing the Replace component to delete the result.