Add TemplateDatabase and remove REMARK parsing
This commit is part of a move away from the storage of template metadata
in REMARK fields of PDB files. Storing arbitrary data in REMARK fields
is a pain: it tends to either violate the 80-character limit of PDB
files or must be painstakingly split into lines, and it adds a lot of
clutter. It doesn't express any information about the relationships
between data, and isn't easily searchable. This will hopefully allow us
to get rid of the silly reduce/expand step we are using at the moment to
load templates with identical sequence.

This commit adds the TemplateDatabase object, which is backed by an
sqlite database. This database stores PDB entries and metadata, as well
as entries for each template (i.e. chain) and template metadata.

To facilitate this, this commit alters several components that treated
the ``structure`` key as a template file containing metadata. These
components now expect a ``template_obj`` key, containing a Template
object.