Refactor configuration extraction
This commit extracts all the logic used for slicing and merging pipelie
configurations out of the Component class and into a PipelineConfig
class. It is now simple to extract slices of the pipeline state using
the PipelineConfig.extract method. It is also now simple to merge slices
of the pipeline config with component parameters using
PipelineConfig.merge_params.