Bugfix: Explicitly convert _exptl.method to array
In rare cases, multiple methods can be used to determine a protein
structure. In these cases, we arbitrarily select the first method to
store in our fold library. This means that the _exptl.method field is
almost always scalar, but can be an array. This bugfix explicitly
converts that field to an array before choosing the first one.