Stop ChainPDBBuilder adding "template_obj" key
The ChainPDBBuilder must be the final component in a Map because it can
return multiple entries for each input and the normal pipelines cannot
cope with that. This can cause big problems if the output from that Map
is pickled, either manually or by the checkpointing mechanism. Use the
BuildTemplate component instead.