Option to do shallow copy for .component.Branch
One of the use cases for `.component.Branch` is to fork a pipeline
before passing a job to a remote machine so that the pipeline can be
cleared of any unpickleable objects. However, Python's implementation of
`copy.deepcopy()` pickles and unpickles the object to be copied.

This commit adds the option to do a shallow copy rather than a deep
copy. This is good enough for the "fork so we can temporarily remove
some keys" case, and the deep copy can be used for the "fork so we can
completely destroy the pipeline for some side effects."