Option to update rather than replace remote data
Often, only a few components of the pipeline state need to be
transferred to a remote machine via a `.pbs.qsub` component. Because the
state must be serialised before it can be transferred, it can be useful
to remove fields from the pipeline state before transferring it to the
remote machine. If this is done in a `.component.Branch` child pipeline,
then the main branch of the pipeline will have extra fields that the
child does not.

This commit adds the option to update the main pipeline with the results
of the child pipelines rather than replacing everything.