Fix ChainPDBBuilder and BuildTemplate
The implementation of BuildTemplate was flawed: it gave the wrong
mapping of original residues, because it just parsed them from the
template PDB file which had already been renumbered.

This commit changes ChainPDBBuilder so that it adds enough metadata to
the pipeline state for BuildTemplate to construct a template without
attempting to infer anything.