Remove exclusion of test packages in setup.py
This commit removes the exclusion of the "phyre_engine.test.*" packages
in the "find_packages" call. Previously, we did not package these
because the expectation was that tests would be called before before
installation. However, Conda sensibly runs tests in a separate
environment, which requires the tests to be installable from the built
package.