Add HTTPMap and start switching to requests lib
There is no reason to resist moving to the "requests" library for
anything requiring HTTP requests: we already have several dependencies,
and we shouldn't be scared of dependencies anyway.

This commit switches StructureRetriever to requests, and adds the
HTTPMap component which will allow components to take advantage of
requests' per-session Keep-Alive.