Add OpenCopy and RestoreSQLDump components
These components should be useful for speeding up the creation of fold
libraries. At the moment, some computation has to be done on the master
to avoid pickling Template objects (which pickle Bio.PDB.Chain objects,
which are huge). This allows the computation to be done on the nodes,
and then the resulting SQL statements to be copied to the master for
execution.